test(results-writer): consolidate three test files into one
Previously the ResultsWriter coverage was split across:

  test_results_writer.py             — component wiring + finalize hook (8)
  test_write_results.py              — schema + missing-tracker (5)
  test_write_results_hierarchical_ids.py — hierarchical-ID semantics (6)

Three files, lots of duplicated scaffolding (scenario factories,
_make_model helpers), and no clear story for a reader trying to find
where a given concern is tested. Merge all 19 into a single
test_results_writer.py organized in 7 sections via comment dividers:

  1. Basic end-to-end (writes, custom path, default path, opt-out, cwd)
  2. Per-tick dispatch contract (NOT in model.phases)
  3. StateTracker discovery (fail-fast, most-granular selection)
  4. JSON schema (full per-group, global-only)
  5. Hierarchical IDs (aggregation_level=0 vs =1, peak_tick)
  6. Robustness (corrupted R, unsorted-scenario axis-vs-label)
  7. finalize() hook coverage (the BaseLaserModel hook ResultsWriter uses)

Drop:
  test_write_results.py
  test_write_results_hierarchical_ids.py

Keep tests/unit/test_tracker_axis_label_alignment.py separate — it
tests the BaseStateTracker / BaseCaseSurveillance label-vs-axis
contract directly, not via ResultsWriter, so it belongs with the
tracker tests when more land.

All 23 tests still pass (19 ResultsWriter + 4 tracker alignment).
