--- tests/unit/test_write_results_hierarchical_ids.py
@@ -1,258 +1,0 @@
-"""Tests for the ResultsWriter component — hierarchical-ID edge cases.
-
-These tests were written because p16 in the laser-mcp prompt suite kept
-failing with broadcast errors like ``(2,) vs (100,)``. The root cause
-was a contract gap between the results JSON and ``StateTracker``:
-
-  - ``two_cluster_scenario`` produces hierarchical IDs of the form
-    ``"cluster_1:node_42"``.
-  - With ``StateTracker(aggregation_level=0)`` the tracker groups by the
-    first segment and yields ``len(group_ids) == 2`` (the two cluster
-    names), not ``n_patches == 100``.
-  - The old schema emitted these length-2 arrays under the misleading
-    name ``summary.attack_rate_per_patch``, while ``num_patches: 100``
-    at the top-level lied about the grouping.
-
-Fix: rename ``_per_patch → _per_group`` (also ``num_patches → num_groups``
-and ``patch_ids → group_ids``) and add ``group_aggregation_level`` so
-consumers can branch on it without re-deriving from the scenario.
-Also ``peak_day → peak_tick`` so the tick index isn't confused with
-calendar days.
-
-These tests pin the new schema. They drive the writer via
-``model.components += [ResultsWriter]`` and read the resulting JSON
-from disk — there is no model.write_results() public API.
-"""
-
-from __future__ import annotations
-
-import json
-
-import polars as pl
-import pytest
-
-from laser.measles.abm import ABMModel
-from laser.measles.abm import ABMParams
-from laser.measles.abm.components import InfectionProcess
-from laser.measles.abm.components import InfectionSeedingParams
-from laser.measles.abm.components import InfectionSeedingProcess
-from laser.measles.abm.components import NoBirthsProcess
-from laser.measles.abm.components import StateTracker
-from laser.measles.biweekly import BiweeklyModel
-from laser.measles.biweekly import BiweeklyParams
-from laser.measles.biweekly.components import InfectionProcess as BWInfection
-from laser.measles.biweekly.components import InitializeEquilibriumStatesProcess
-from laser.measles.components import BaseStateTrackerParams
-from laser.measles.components import ResultsWriter
-from laser.measles.components import ResultsWriterParams
-from laser.measles.components import create_component
-
-N_PER_CLUSTER = 4
-TICKS = 60
-
-
-def _hierarchical_scenario(n_per_cluster: int = N_PER_CLUSTER) -> pl.DataFrame:
-    """Two-cluster scenario with ``cluster_X:node_Y``-style hierarchical IDs.
-
-    Mirrors the shape of ``laser.measles.scenarios.two_cluster_scenario``
-    but is built inline so the test stays self-contained.
-    """
-    ids: list[str] = []
-    lats: list[float] = []
-    lons: list[float] = []
-    pops: list[int] = []
-    for cluster_idx in range(2):
-        for i in range(n_per_cluster):
-            ids.append(f"cluster_{cluster_idx + 1}:node_{i + 1}")
-            lats.append(40.0 + cluster_idx * 5.0 + i * 0.01)
-            lons.append(5.0 + cluster_idx * 5.0 + i * 0.01)
-            pops.append(50_000 + 1_000 * i)
-    return pl.DataFrame(
-        {
-            "id": ids,
-            "pop": pops,
-            "lat": lats,
-            "lon": lons,
-            "mcv1": [0.0] * len(ids),
-        }
-    )
-
-
-def _build_and_run(state_tracker_params, out_path) -> dict:
-    """Run a small ABM with the given tracker config + ResultsWriter,
-    then return the JSON the writer produced."""
-    params = ABMParams(num_ticks=TICKS, seed=42, start_time="2000-01", verbose=False, show_progress=False)
-    model = ABMModel(_hierarchical_scenario(), params)
-    model.add_component(NoBirthsProcess)
-    model.add_component(create_component(InfectionSeedingProcess, params=InfectionSeedingParams(num_infections=20)))
-    model.add_component(InfectionProcess)
-    model.add_component(create_component(StateTracker, params=state_tracker_params))
-    model.add_component(create_component(ResultsWriter, params=ResultsWriterParams(path=str(out_path))))
-    model.run()
-    return json.loads(out_path.read_text())
-
-
-def test_aggregation_level_0_with_hierarchical_ids_rolls_up_to_clusters(tmp_path):
-    """With ``aggregation_level=0`` and hierarchical IDs (``cluster_1:node_1``),
-    the tracker rolls nodes up to one row per top-level segment.
-    ResultsWriter then emits length-2 arrays — but under the
-    correctly-named ``_per_group`` keys, with ``num_groups=2`` and
-    ``group_aggregation_level=0`` so the consumer knows what they're
-    holding.
-    """
-    n_clusters = 2
-    out = _build_and_run(BaseStateTrackerParams(aggregation_level=0), tmp_path / "results.json")
-    summary = out["summary"]
-
-    assert out["num_groups"] == n_clusters
-    assert out["group_aggregation_level"] == 0
-    assert sorted(out["group_ids"]) == ["cluster_1", "cluster_2"]
-
-    assert summary["attack_rate_per_group"] is not None
-    assert len(summary["attack_rate_per_group"]) == n_clusters
-    assert summary["peak_infectious_per_group"] is not None
-    assert len(summary["peak_infectious_per_group"]) == n_clusters
-    for state, vec in summary["final_state_per_group"].items():
-        assert len(vec) == n_clusters, f"final_state_per_group[{state!r}] wrong length"
-
-
-def test_aggregation_level_1_with_hierarchical_ids_is_true_per_patch(tmp_path):
-    """When ``aggregation_level`` is set to the ID depth minus one, the
-    tracker yields one row per scenario patch and ResultsWriter
-    must produce per-group arrays of that exact length — which IS
-    per-patch in this case.
-    """
-    n_patches = 2 * N_PER_CLUSTER
-    out = _build_and_run(BaseStateTrackerParams(aggregation_level=1), tmp_path / "results.json")
-    summary = out["summary"]
-
-    assert out["num_groups"] == n_patches
-    assert out["group_aggregation_level"] == 1
-    assert sorted(out["group_ids"]) == sorted(f"cluster_{c}:node_{i + 1}" for c in (1, 2) for i in range(N_PER_CLUSTER))
-    assert summary["attack_rate_per_group"] is not None
-    assert len(summary["attack_rate_per_group"]) == n_patches
-    assert summary["peak_infectious_per_group"] is not None
-    assert len(summary["peak_infectious_per_group"]) == n_patches
-    for state, vec in summary["final_state_per_group"].items():
-        assert len(vec) == n_patches, f"final_state_per_group[{state!r}] wrong length"
-
-
-def test_top_level_exposes_aggregation_level_for_disambiguation(tmp_path):
-    """Consumers reading the JSON must be able to tell whether the
-    ``_per_group`` arrays are at patch granularity or aggregated above.
-    The top-level ``group_aggregation_level`` field carries that
-    contract so readers can branch without re-deriving from the
-    scenario.
-    """
-    out = _build_and_run(BaseStateTrackerParams(aggregation_level=0), tmp_path / "results.json")
-    assert "group_aggregation_level" in out
-    assert isinstance(out["group_aggregation_level"], int)
-    assert out["group_aggregation_level"] == 0
-
-
-def test_peak_tick_is_tick_index_in_biweekly_model(tmp_path):
-    """``summary.peak_tick`` is named for what it is — the tick index
-    of the global infectious peak — so consumers don't accidentally
-    treat a biweekly tick (14 days) as a calendar day.
-    """
-    out_path = tmp_path / "results.json"
-    scenario = pl.DataFrame(
-        {
-            "id": [f"patch_{i}" for i in range(3)],
-            "pop": [10_000, 20_000, 15_000],
-            "lat": [0.0, 0.1, 0.2],
-            "lon": [0.0, 0.0, 0.0],
-            "mcv1": [0.0, 0.0, 0.0],
-        }
-    )
-    params = BiweeklyParams(num_ticks=52, seed=42, start_time="2000-01", verbose=False, show_progress=False)
-    model = BiweeklyModel(scenario, params)
-    model.components = [
-        InitializeEquilibriumStatesProcess,
-        BWInfection,
-        create_component(StateTracker, params=BaseStateTrackerParams(aggregation_level=0)),
-        create_component(ResultsWriter, params=ResultsWriterParams(path=str(out_path))),
-    ]
-    model.run()
-    out = json.loads(out_path.read_text())
-    summary = out["summary"]
-
-    assert "peak_tick" in summary
-    assert "peak_day" not in summary, "peak_day was renamed to peak_tick to avoid implying calendar days."
-    assert 0 <= summary["peak_tick"] < params.num_ticks
-
-
-def test_results_writer_path_is_relative_to_cwd(tmp_path, monkeypatch):
-    """Verify the writer resolves a relative ``path`` against the
-    current working directory. The laser-mcp parallel test runner
-    relies on this for per-attempt cwd isolation — if the writer ever
-    starts resolving relative to the package or model directory,
-    concurrent prompts will overwrite each other's results.json.
-    """
-    monkeypatch.chdir(tmp_path)
-
-    params = ABMParams(num_ticks=TICKS, seed=42, start_time="2000-01", verbose=False, show_progress=False)
-    model = ABMModel(_hierarchical_scenario(), params)
-    model.add_component(NoBirthsProcess)
-    model.add_component(create_component(InfectionSeedingProcess, params=InfectionSeedingParams(num_infections=20)))
-    model.add_component(InfectionProcess)
-    model.add_component(create_component(StateTracker, params=BaseStateTrackerParams(aggregation_level=1)))
-    model.add_component(ResultsWriter)  # default path = "results.json"
-    model.run()
-
-    target = tmp_path / "results.json"
-    assert target.exists(), f"ResultsWriter should write its default path 'results.json' relative to cwd ({tmp_path}); found nothing there."
-
-
-def test_group_ids_match_axis_order_for_unsorted_scenario(tmp_path):
-    """Regression for the BaseStateTracker contract: ``group_ids[i]`` must
-    correspond to ``state_tracker[..., i]``. Previously ``group_ids`` was
-    sorted while the tracker's last axis is filled in the insertion order
-    of ``node_mapping.items()`` — so for a scenario whose IDs aren't already
-    in sorted order, the labels in the emitted JSON were misaligned with
-    the per-group arrays.
-
-    Forces an unsorted ID order at scenario construction; verifies the
-    emitted ``group_ids`` reflects scenario row order AND the per-group
-    arrays are consistent with that order (the patch with the highest
-    initial pop should also have the largest peak_infectious_per_group).
-    """
-    out_path = tmp_path / "results.json"
-
-    # IDs chosen so sorted order (apple, banana, cherry) differs from
-    # the scenario row order (banana, apple, cherry).
-    scenario = pl.DataFrame(
-        {
-            "id": ["banana", "apple", "cherry"],
-            "pop": [200_000, 50_000, 80_000],  # banana is largest
-            "lat": [0.0, 0.1, 0.2],
-            "lon": [0.0, 0.0, 0.0],
-            "mcv1": [0.0, 0.0, 0.0],
-        }
-    )
-    params = ABMParams(num_ticks=60, seed=42, start_time="2000-01", verbose=False, show_progress=False)
-    model = ABMModel(scenario, params)
-    model.add_component(NoBirthsProcess)
-    model.add_component(create_component(InfectionSeedingProcess, params=InfectionSeedingParams(num_infections=20)))
-    model.add_component(InfectionProcess)
-    model.add_component(create_component(StateTracker, params=BaseStateTrackerParams(aggregation_level=0)))
-    model.add_component(create_component(ResultsWriter, params=ResultsWriterParams(path=str(out_path))))
-    model.run()
-
-    out = json.loads(out_path.read_text())
-    assert out["group_ids"] == ["banana", "apple", "cherry"], (
-        f"group_ids should reflect scenario row order, not sorted; got {out['group_ids']}"
-    )
-
-    # Sanity: the per-group arrays line up with group_ids. The banana
-    # patch (200K pop, index 0) should have a strictly larger peak than
-    # apple (50K, index 1) — would be impossible if labels and array
-    # were misaligned (apple would appear at index 0).
-    peaks = out["summary"]["peak_infectious_per_group"]
-    assert peaks[0] > peaks[1], (
-        f"index 0 should be banana (largest pop), but peak[0]={peaks[0]} < peak[1]={peaks[1]} — labels and axis are out of sync"
-    )
-
-
-if __name__ == "__main__":
-    pytest.main([__file__, "-v"])
